Added plotting facilities, renamed binaries
cut-flow-plots creates a set of paneled plots showing the cuts applied
annotated with their names.

Added AtlasStyle
Added python and a small script that was the first attempt at
cut-flow-plots, this is historical interest only since PyROOT's memory
management sucks.

Renamed main.cxx to skim-tree.cxx (and the resulting binary to match
this). This is to reflect the code's intent, which is to apply cuts to a
subset (ie skim) of the tree.